Tail from multiple control towers for higher availability per log
Summary:
All copilot subscriptions go to multiple control towers. If any single control tower goes down, it should not longer affect the read availability of any log.

I removed code that reconnects to control towers on goodbye because it was unthrottled and causing problems. Will update once we have Relja's timers.

The subscription to 0 bug is still there. Will fix in next diff (wanted to do it separately).

Test Plan:
make check
(also ran benchmarks)

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: eduardosuarez, dhruba

Differential Revision: https://reviews.facebook.net/D38937

+++ src/copilot/options.h
@@ -77,7 +77,7 @@ struct CopilotOptions {
   uint32_t consistent_hash_replicas;
 
   // Each topic should be tailed by this many control towers.
-  uint32_t control_towers_per_log;
+  size_t control_towers_per_log;
 
   // Number of connections between this copilot and each control tower.
   // Should be <= number of message loop workers.
